[ANDROSDK-1701] Refactor TrackedEntityEnrollmentOrphanCleaner class
